Moved string utilities to opm/common
