+++ src/opm/parser/eclipse/EclipseState/EndpointScaling.cpp
@@ -1,6 +1,6 @@
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 #include <opm/parser/eclipse/EclipseState/EndpointScaling.hpp>
-#include <opm/parser/eclipse/Utility/String.hpp>
+#include <opm/common/utility/String.hpp>
 
 namespace Opm {
