+++ src/opm/parser/eclipse/EclipseState/Schedule/TimeMap.cpp
@@ -32,7 +32,7 @@
 #include <opm/parser/eclipse/Deck/DeckKeyword.hpp>
 #include <opm/parser/eclipse/Deck/DeckRecord.hpp>
 #include <opm/parser/eclipse/EclipseState/Schedule/TimeMap.hpp>
-#include <opm/parser/eclipse/Utility/String.hpp>
+#include <opm/common/utility/String.hpp>
 
 
 constexpr const std::time_t invalid_time = -1;
